Add a MonadCatchIO instance for WebSocket p
This version does it the lazy way: it imports Snap.Iteratee from
snap-core. Snap.Iteratee defines a MonadCatchIO instance for
Iteratee, and with that in hand we can just derive MonadCatchIO.

snap-core is a fairly heavy dependency, so a better approach
might be to steal the Iterattee instance, rather than adding the
dependency.